move the functions used in the toolbar to the toolbar class
